requirements.txt
annotated-types==0.7.0
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1761591974641/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1759648874697/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1761202849623/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1760437218288/work
click==8.3.0
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1759948064277/work
gitdb==4.0.12
GitPython==3.1.45
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1756739574751/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1760286409563/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1759055168201/work
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
networkx @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_networkx_1749078300/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_numpy_1761162089/work/dist/numpy-2.3.4-cp311-cp311-linux_x86_64.whl#sha256=d9bbdcee86ff2d477e5b58162d14e441c7693336c490a49c1fcacc11ee1b868e
packaging==25.0
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1675487189013/work
platformdirs==4.5.0
protobuf==6.33.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.12.3
pydantic_core==2.41.4
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1758891829620/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1755614211359/work
sentry-sdk==2.43.0
smmap==5.0.2
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1745946051654/work
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
triton==3.1.0
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
wandb==0.22.3
zstandard==0.25.0
